Added automatic tests for stim map and inverse stim map routines
